Split mapper selection state
