+++ Sources/KeyPathAppKit/UI/Experimental/MapperViewModel.swift
@@ -268,388 +268,6 @@ class MapperViewModel {
     /// Delay before finalizing a sequence capture (allows for multi-key sequences)
     @ObservationIgnored let sequenceFinalizeDelay: TimeInterval = 0.8
 
-    var canSave: Bool {
-        inputSequence != nil && (outputSequence != nil || selectedApp != nil || selectedSystemAction != nil || selectedURL != nil)
-    }
-
-    var hasShiftedOutputConfigured: Bool {
-        shiftedOutputSequence != nil
-    }
-
-    /// Whether the current input key has a standard shift symbol (e.g., 1→!, ;→:)
-    var defaultShiftSymbol: String? {
-        LabelMetadata.forLabel(inputLabel).shiftSymbol
-    }
-
-    var shiftedOutputBlockingReason: String? {
-        if selectedAppCondition != nil {
-            return "Shift output is only available for rules that apply everywhere."
-        }
-        if selectedApp != nil || selectedSystemAction != nil || selectedURL != nil {
-            return "Shift output works only with keystroke output."
-        }
-        if advancedBehavior.hasAdvancedConfig {
-            return "Shift output isn't available with hold, combo, or multi-tap behaviors."
-        }
-        return nil
-    }
-
-    var canUseShiftedOutput: Bool {
-        shiftedOutputBlockingReason == nil
-    }
-
-    var isIdentityKeystrokeMapping: Bool {
-        guard selectedApp == nil,
-              selectedSystemAction == nil,
-              selectedURL == nil,
-              !advancedBehavior.hasAdvancedConfig,
-              let inputKanata = currentInputKanataString(),
-              let outputSequence
-        else {
-            return false
-        }
-
-        let outputKanata = convertSequenceToKanataFormat(outputSequence)
-        return inputKanata.caseInsensitiveCompare(outputKanata) == .orderedSame
-    }
-
-    func configure(kanataManager: RuntimeCoordinator) {
-        self.kanataManager = kanataManager
-        Task {
-            await refreshAvailableLayers()
-        }
-    }
-
-    /// Set the current layer
-    func setLayer(_ layer: String) {
-        currentLayer = layer
-        AppLogger.shared.log("🗂️ [MapperViewModel] Layer set to: \(layer)")
-
-        // Post notification for other views to update
-        NotificationCenter.default.post(
-            name: .kanataLayerChanged,
-            object: nil,
-            userInfo: ["layer": layer]
-        )
-    }
-
-    /// Update input from a key click in the overlay (used by mapper drawer)
-    func setInputFromKeyClick(
-        keyCode: UInt16,
-        inputLabel: String,
-        outputLabel: String,
-        appIdentifier: String? = nil,
-        systemActionIdentifier: String? = nil,
-        urlIdentifier: String? = nil,
-        shiftedOutputKey: String? = nil
-    ) {
-        // Stop any active recording
-        stopRecording()
-
-        // Update input
-        inputKeyCode = keyCode
-        self.inputLabel = formatKeyForDisplay(inputLabel)
-        inputSequence = KeySequence(
-            keys: [KeyPress(baseKey: inputLabel, modifiers: [], keyCode: Int64(keyCode))],
-            captureMode: .single
-        )
-
-        // Clear previous selections first (including app/device conditions - revert to "Everywhere")
-        selectedApp = nil
-        selectedSystemAction = nil
-        selectedURL = nil
-        selectedAppCondition = nil
-        selectedDeviceCondition = nil
-        clearShiftedOutput()
-
-        // Set output based on action type
-        if let appId = appIdentifier, let appInfo = appLaunchInfo(for: appId) {
-            selectedApp = appInfo
-            self.outputLabel = appInfo.name
-            outputSequence = nil
-            AppLogger.shared.log("🖱️ [MapperViewModel] Key click - app launch: \(inputLabel) -> \(appInfo.name)")
-        } else if let urlId = urlIdentifier {
-            selectedURL = urlId
-            self.outputLabel = extractDomain(from: urlId)
-            outputSequence = nil
-            AppLogger.shared.log("🖱️ [MapperViewModel] Key click - URL: \(inputLabel) -> \(urlId)")
-        } else if let systemId = systemActionIdentifier, let systemAction = SystemActionInfo.find(byOutput: systemId) {
-            selectedSystemAction = systemAction
-            self.outputLabel = systemAction.name
-            outputSequence = nil
-            AppLogger.shared.log("🖱️ [MapperViewModel] Key click - system action: \(inputLabel) -> \(systemAction.name)")
-        } else if let systemAction = SystemActionInfo.find(byOutput: outputLabel) {
-            // Fallback path: some callers pass only a display label (e.g., "Notification Center")
-            // before the layer metadata has a concrete systemActionIdentifier.
-            selectedSystemAction = systemAction
-            self.outputLabel = systemAction.name
-            outputSequence = nil
-            AppLogger.shared.log("🖱️ [MapperViewModel] Key click - system action (fallback): \(inputLabel) -> \(systemAction.name)")
-        } else {
-            // Regular key mapping
-            self.outputLabel = formatKeyForDisplay(outputLabel)
-            outputSequence = KeySequence(
-                keys: [KeyPress(baseKey: outputLabel, modifiers: [], keyCode: 0)],
-                captureMode: .single
-            )
-            AppLogger.shared.log("🖱️ [MapperViewModel] Key click - key mapping: \(inputLabel) -> \(outputLabel)")
-        }
-        applyShiftedOutputPreset(shiftedOutputKey)
-
-        // If no custom shifted output was loaded, pre-populate with system default
-        if !hasShiftedOutputConfigured, let defaultShift = defaultShiftSymbol {
-            shiftedOutputLabel = defaultShift
-            shiftedOutputSequence = KeySequence(
-                keys: [KeyPress(baseKey: defaultShift, modifiers: [], keyCode: 0)],
-                captureMode: .single
-            )
-            isShiftedOutputDefault = true
-        }
-
-        // Update list of apps that have mappings for this key
-        Task { await updateAppsWithMapping() }
-
-        // Clear previous behavior (will be loaded separately via loadBehaviorFromExistingRule)
-        advancedBehavior.holdAction = ""
-        advancedBehavior.doubleTapAction = ""
-        advancedBehavior.macroBehavior = nil
-        advancedBehavior.comboKeys = []
-        advancedBehavior.comboOutput = ""
-        advancedBehavior.holdBehavior = .basic
-        advancedBehavior.tapTimeout = 200
-        advancedBehavior.holdTimeout = 200
-        advancedBehavior.customTapKeysText = ""
-    }
-
-    /// Load behavior from existing custom rule for the current input key
-    /// Call this after setInputFromKeyClick to restore hold/tap-dance settings
-    func loadBehaviorFromExistingRule(kanataManager: RuntimeCoordinator) {
-        guard let keyCode = inputKeyCode else { return }
-        let inputKey = OverlayKeyboardView.keyCodeToKanataName(keyCode)
-
-        // Look up existing rule
-        guard let existingRule = kanataManager.getCustomRule(forInput: inputKey) else {
-            // No custom rule — populate default shift symbol if available
-            if !hasShiftedOutputConfigured, let defaultShift = defaultShiftSymbol {
-                shiftedOutputLabel = defaultShift
-                shiftedOutputSequence = KeySequence(
-                    keys: [KeyPress(baseKey: defaultShift, modifiers: [], keyCode: 0)],
-                    captureMode: .single
-                )
-                isShiftedOutputDefault = true
-            }
-            AppLogger.shared.log("📖 [MapperViewModel] No existing behavior for input '\(inputKey)'")
-            return
-        }
-        applyShiftedOutputPreset(existingRule.shiftedOutput)
-        // If a custom shifted output was loaded, mark it as non-default
-        if hasShiftedOutputConfigured {
-            isShiftedOutputDefault = false
-        } else if let defaultShift = defaultShiftSymbol {
-            // No custom shift in rule — populate default
-            shiftedOutputLabel = defaultShift
-            shiftedOutputSequence = KeySequence(
-                keys: [KeyPress(baseKey: defaultShift, modifiers: [], keyCode: 0)],
-                captureMode: .single
-            )
-            isShiftedOutputDefault = true
-        }
-
-        guard let behavior = existingRule.behavior else {
-            AppLogger.shared.log("📖 [MapperViewModel] No existing behavior for input '\(inputKey)'")
-            return
-        }
-
-        switch behavior {
-        case let .dualRole(dualRole):
-            advancedBehavior.holdAction = dualRole.holdActionString
-            advancedBehavior.tapTimeout = dualRole.tapTimeout
-            advancedBehavior.holdTimeout = dualRole.holdTimeout
-
-            // Determine hold behavior type from flags
-            if dualRole.activateHoldOnOtherKey {
-                advancedBehavior.holdBehavior = .triggerEarly
-            } else if dualRole.quickTap {
-                advancedBehavior.holdBehavior = .quickTap
-            } else if !dualRole.customTapKeys.isEmpty {
-                advancedBehavior.holdBehavior = .customKeys
-                advancedBehavior.customTapKeysText = dualRole.customTapKeys.joined(separator: " ")
-            } else {
-                advancedBehavior.holdBehavior = .basic
-            }
-
-            AppLogger.shared.log("📖 [MapperViewModel] Loaded dualRole behavior for '\(inputKey)': hold='\(dualRole.holdAction)'")
-
-        case let .tapOrTapDance(tapBehavior):
-            switch tapBehavior {
-            case .tap:
-                AppLogger.shared.log("📖 [MapperViewModel] Loaded tap behavior for '\(inputKey)'")
-            case let .tapDance(tapDance):
-                // Load timing
-                advancedBehavior.tapTimeout = tapDance.windowMs
-
-                // Load tap-dance steps
-                // First step (index 0) is single tap - already loaded in outputLabel
-                if tapDance.steps.count > 1 {
-                    advancedBehavior.doubleTapAction = tapDance.steps[1].actionString
-                }
-                // Load additional steps into tapDanceSteps array (triple tap, quad tap, etc.)
-                if tapDance.steps.count > 2 {
-                    advancedBehavior.tapDanceSteps = tapDance.steps.dropFirst(2).map { step in
-                        (label: step.label, action: step.actionString, isRecording: false)
-                    }
-                }
-
-                AppLogger.shared.log("📖 [MapperViewModel] Loaded tapDance behavior for '\(inputKey)': \(tapDance.steps.count) steps, windowMs=\(tapDance.windowMs)")
-            }
-
-        case let .macro(macro):
-            advancedBehavior.macroBehavior = macro
-            AppLogger.shared.log("📖 [MapperViewModel] Loaded macro behavior for '\(inputKey)'")
-
-        case let .chord(chord):
-            // Load chord: remove the input key from the keys list (it's implicit)
-            advancedBehavior.comboKeys = chord.keys.filter { $0.lowercased() != inputKey.lowercased() }
-            advancedBehavior.comboOutput = chord.outputString
-            advancedBehavior.comboTimeout = chord.timeout
-
-            AppLogger.shared.log("📖 [MapperViewModel] Loaded chord behavior for '\(inputKey)': keys=\(chord.keys), output='\(chord.outputString)'")
-        }
-    }
-
-    /// Update the list of apps that have a mapping for the currently selected input key
-    func updateAppsWithMapping() async {
-        guard let keyCode = inputKeyCode else {
-            appsWithCurrentKeyMapping = []
-            return
-        }
-        let inputKey = OverlayKeyboardView.keyCodeToKanataName(keyCode)
-        appsWithCurrentKeyMapping = await AppKeymapStore.shared.getAppsWithMapping(forInputKey: inputKey)
-    }
-
-    /// Apply preset values from overlay click
-    func applyPresets(
-        input: String,
-        output: String,
-        layer: String? = nil,
-        inputKeyCode: UInt16? = nil,
-        appIdentifier: String? = nil,
-        systemActionIdentifier: String? = nil,
-        urlIdentifier: String? = nil,
-        shiftedOutput: String? = nil
-    ) {
-        // Stop any active recording
-        stopRecording()
-
-        // Store original context for reset after clear
-        originalInputKey = input
-        originalOutputKey = output
-        originalShiftedOutputKey = shiftedOutput
-        originalAppIdentifier = appIdentifier
-        originalSystemActionIdentifier = systemActionIdentifier
-        originalURL = urlIdentifier
-        originalLayer = layer
-
-        // Clear any previously saved rule ID since we're starting fresh
-        lastSavedRuleID = nil
-        selectedApp = nil
-        selectedSystemAction = nil
-        selectedURL = nil
-        clearShiftedOutput()
-
-        // Set the layer
-        if let layer {
-            currentLayer = layer
-        }
-
-        // Set the input label and sequence
-        inputLabel = formatKeyForDisplay(input)
-        // Create simple key sequences for the presets
-        // Use provided keyCode if available (from overlay), otherwise 0 as placeholder
-        let keyCodeToUse = inputKeyCode ?? 0
-        inputSequence = KeySequence(
-            keys: [KeyPress(baseKey: input, modifiers: [], keyCode: Int64(keyCodeToUse))],
-            captureMode: .single
-        )
-
-        if let appIdentifier, let appInfo = appLaunchInfo(for: appIdentifier) {
-            selectedApp = appInfo
-            outputLabel = appInfo.name
-            outputSequence = nil
-            AppLogger.shared.log("🗺️ [MapperViewModel] Preset output is app launch: \(appInfo.name)")
-        } else if let urlIdentifier {
-            selectedURL = urlIdentifier
-            outputLabel = extractDomain(from: urlIdentifier)
-            outputSequence = nil
-            AppLogger.shared.log("🗺️ [MapperViewModel] Preset output is URL: \(urlIdentifier)")
-        } else if let systemActionIdentifier,
-                  let systemAction = SystemActionInfo.find(byOutput: systemActionIdentifier)
-        {
-            // It's a system action/media key - set selectedSystemAction for SF Symbol rendering
-            selectedSystemAction = systemAction
-            outputLabel = systemAction.name
-            outputSequence = nil
-            AppLogger.shared.log("🗺️ [MapperViewModel] Preset output is system action: \(systemAction.name)")
-        } else if let systemAction = SystemActionInfo.find(byOutput: output) {
-            // Fallback: resolve by output label
-            selectedSystemAction = systemAction
-            outputLabel = systemAction.name
-            outputSequence = nil
-            AppLogger.shared.log("🗺️ [MapperViewModel] Preset output is system action: \(systemAction.name)")
-        } else {
-            // Regular key mapping
-            outputLabel = formatKeyForDisplay(output)
-            outputSequence = KeySequence(
-                keys: [KeyPress(baseKey: output, modifiers: [], keyCode: 0)],
-                captureMode: .single
-            )
-        }
-        applyShiftedOutputPreset(shiftedOutput)
-
-        // Store the keyCode for proper keycap rendering
-        if let inputKeyCode {
-            self.inputKeyCode = inputKeyCode
-        }
-
-        statusMessage = nil
-        statusIsError = false
-
-        AppLogger.shared.log("📝 [MapperViewModel] Applied presets: \(input) → \(output) [layer: \(currentLayer)] [inputKeyCode: \(keyCodeToUse)]")
-    }
-
-    /// Format a kanata key name for display (e.g., "leftmeta" -> "⌘")
-    /// Uses the centralized KeyDisplayFormatter utility.
-    func formatKeyForDisplay(_ key: String) -> String {
-        AppLogger.shared.log("🔤 [MapperViewModel] formatKeyForDisplay input: '\(key)'")
-        let result = KeyDisplayFormatter.format(key)
-        AppLogger.shared.log("🔤 [MapperViewModel] formatKeyForDisplay output: '\(result)'")
-        return result
-    }
-
-    func formattedSequenceForDisplay(_ sequence: String) -> String {
-        sequence
-            .split(separator: " ")
-            .map { formatKeyForDisplay(String($0)) }
-            .joined(separator: " ")
-    }
-
-    func applyShiftedOutputPreset(_ shiftedOutput: String?) {
-        guard let shiftedOutput = shiftedOutput?.trimmingCharacters(in: .whitespacesAndNewlines),
-              !shiftedOutput.isEmpty
-        else {
-            clearShiftedOutput()
-            return
-        }
-
-        originalShiftedOutputKey = shiftedOutput
-        shiftedOutputLabel = formattedSequenceForDisplay(shiftedOutput)
-        shiftedOutputSequence = KeySequence(
-            keys: [KeyPress(baseKey: shiftedOutput, modifiers: [], keyCode: 0)],
-            captureMode: .single
-        )
-        isShiftedOutputDefault = false
-    }
-
     /// Show a success status message that auto-dismisses after a delay
     /// Show a success status message that auto-dismisses after a delay
     func showTransientStatus(_ message: String, duration: Duration = .seconds(2)) {
